backend/requirements.txt

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic[email]>=2.10.0
aiosmtplib>=3.0.0

